fix(cz-commitlint): fix minor formatting issues
This commit adds a space after the title before the skip message and removes the double colon in the
`config-conventional` type of change prompt. For example (when using
`@commitlint/config-conventional`), the following would previously be printed:

```
Select the type of change that you're committing::  feat
What is the scope of this change (e.g. component or file name)(press enter to skip): (max 96 chars)
```

Now with this change:

```
Select the type of change that you're committing: feat
What is the scope of this change (e.g. component or file name) (press enter to skip): (max 96 chars)
```